# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sigvitas HRMS API",
    "version": "1"
  },
  "paths": {
...
    "/api/v1/settings/{key}": {
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Change Setting",
        "description": "`security.*` keys additionally need `security.settings.manage`; both need step-up.",
        "operationId": "change_setting_api_v1_settings__key__put",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 100,
              "title": "Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingChange"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SettingChange": {
        "properties": {
          "value": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "value"
        ],
        "title": "SettingChange"
      },
...
      "SettingView": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "value": {
            "type": "integer",
            "title": "Value"
          },
          "default": {
            "type": "integer",
            "title": "Default"
          },
          "minimum": {
            "type": "integer",
            "title": "Minimum"
          },
          "maximum": {
            "type": "integer",
            "title": "Maximum"
          },
          "unit": {
            "type": "string",
            "enum": [
              "attempts",
              "minutes",
              "hours",
              "days"
            ],
            "title": "Unit"
          },
          "overridden": {
            "type": "boolean",
            "title": "Overridden"
          },
          "permission": {
            "type": "string",
            "enum": [
              "security.settings.manage",
              "settings.manage"
            ],
            "title": "Permission"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "description",
          "value",
          "default",
          "minimum",
          "maximum",
          "unit",
          "overridden",
          "permission"
        ],
        "title": "SettingView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}